reorginize api messages
